resume-analyzer-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
python-dotenv==1.0.0




pydantic==2.5.2
PyMuPDF==1.23.8
docx2txt==0.8
google-generativeai==0.3.2
Pillow==10.2.0